[Migration] Restructure rewriter core under onnxscript
- Moving files to new directories without changing.
- Merging ir/__init__.py with existing ir/__init__.py. The two IRs are expected to be merged in the future.

[ghstack-poisoned]
