+++ onnxscript/optimizer/constant_folding.py
