+++ onnxscript/optimizer/fold_constants_v0.py
