+++ onnxscript/optimizer/function_folding_test.py
