+++ onnxscript/optimizer/remove_unused_function.py
