+++ onnxscript/optimizer/simple_function_folding.py
