+++ onnxscript/optimizer/simple_function_folding_test.py
