+++ onnxscript/rewriter/onnxruntime/softmax.py
